Enabled the use of Multiple SSO plugins.
The Shibboleth SSO plugin and documentation was updated to enable the use of multiple SSOs.
The issue was that to enable shibboleth authentication the path /redbox/default/sso was
being placed behind shibboleth authentication. This meant that shibboleth would interfere
with the other SSO plugins. To fix this the interface links to:

    /redbox/default/sso/shibboleth?ssoID=Shibboleth

instead of

    /redbox/default/sso?ssoID=Shibboleth

That the path /redbox/default/sso/shibboleth doesn't exist doesn't matter, as the ssoID=Shibboleth
triggers the PortalSecurityManager to redirect to the url given by the SSO's ssoGetRemoteLogonURL
method.